packages/autofix/requirements.txt
flask==3.1.3
requests==2.32.5
Django==5.2.11
google-genai==1.64.0
python-dotenv==1.2.1
